+++ lib/elixir/lib/record.ex
@@ -32,7 +32,7 @@ defmodule Record do
         Record.defrecord :user, name: "john", age: 25
 
         @type user :: record(:user, name: String.t, age: integer)
-        # expands to: `@type user :: {:user, String.t, integer}`
+        # expands to: "@type user :: {:user, String.t, integer}"
       end
   """