+++ lib/mix/lib/mix/tasks/deps.compile.ex
@@ -63,8 +63,8 @@ defmodule Mix.Tasks.Deps.Compile do
           make?(dep) ->
             do_make dep, config
           true ->
-            shell.error "Could not compile #{app}, no mix.exs, rebar.config or Makefile " <>
-              "(pass :compile as an option to customize compilation, set it to false to do nothing)"
+            shell.error "Could not compile #{inspect app}, no \"mix.exs\", \"rebar.config\" or \"Makefile\" " <>
+              "(pass :compile as an option to customize compilation, set it to \"false\" to do nothing)"
         end
 
         unless mix?(dep), do: build_structure(dep, config)
@@ -83,8 +83,8 @@ defmodule Mix.Tasks.Deps.Compile do
   end
 
   defp check_unavailable!(app, {:unavailable, _}) do
-    Mix.raise "Cannot compile dependency #{app} because " <>
-      "it isn't available, run `mix deps.get` first"
+    Mix.raise "Cannot compile dependency #{inspect app} because " <>
+      "it isn't available, run \"mix deps.get\" first"
   end
 
   defp check_unavailable!(_, _) do
@@ -94,7 +94,7 @@ defmodule Mix.Tasks.Deps.Compile do
   defp do_mix(dep, _config) do
     Mix.Dep.in_dependency dep, fn _ ->
       if req = old_elixir_req(Mix.Project.config) do
-        Mix.shell.error "warning: the dependency #{dep.app} requires Elixir #{inspect req} " <>
+        Mix.shell.error "warning: the dependency #{inspect dep.app} requires Elixir #{inspect req} " <>
                         "but you are running on v#{System.version}"
       end
 
@@ -105,9 +105,9 @@ defmodule Mix.Tasks.Deps.Compile do
         kind, reason ->
           stacktrace = System.stacktrace
           app = dep.app
-          Mix.shell.error "could not compile dependency #{app}, mix compile failed. " <>
-            "You can recompile this dependency with `mix deps.compile #{app}` or " <>
-            "update it with `mix deps.update #{app}`"
+          Mix.shell.error "could not compile dependency #{inspect app}, \"mix compile\" failed. " <>
+            "You can recompile this dependency with \"mix deps.compile #{app}\" or " <>
+            "update it with \"mix deps.update #{app}\""
           :erlang.raise(kind, reason, stacktrace)
       end
     end
@@ -125,16 +125,16 @@ defmodule Mix.Tasks.Deps.Compile do
 
   defp handle_rebar_not_found(app) do
     shell = Mix.shell
-    shell.info "Could not find rebar, which is needed to build dependency #{inspect app}"
-    shell.info "I can install a local copy which is just used by mix"
+    shell.info "Could not find \"rebar\", which is needed to build dependency #{inspect String.to_atom(app)}"
+    shell.info "I can install a local copy which is just used by Mix"
 
     unless shell.yes?("Shall I install rebar?") do
-      Mix.raise "Could not find rebar to compile " <>
-        "dependency #{app}, please ensure rebar is available"
+      Mix.raise "Could not find \"rebar\" to compile " <>
+        "dependency #{inspect String.to_atom(app)}, please ensure \"rebar\" is available"
     end
 
     (Mix.Tasks.Local.Rebar.run([]) && Mix.Rebar.local_rebar_cmd) ||
-      Mix.raise "rebar installation failed"
+      Mix.raise "\"rebar\" installation failed"
   end
 
   defp do_make(dep, config) do
@@ -158,8 +158,8 @@ defmodule Mix.Tasks.Deps.Compile do
     Mix.Dep.in_dependency dep, fn _ ->
       env = [{"ERL_LIBS", Path.join(config[:build_path], "lib")}]
       if Mix.shell.cmd("#{command} #{extra}", print_app: print_app?, env: env) != 0 do
-        Mix.raise "Could not compile dependency #{app}, #{command} command failed. " <>
-          "If you want to recompile this dependency, please run: mix deps.compile #{app}"
+        Mix.raise "Could not compile dependency #{inspect String.to_atom(app)}, \"#{command}\" command failed. " <>
+          "If you want to recompile this dependency, please run: \"mix deps.compile #{app}\""
       end
     end
     true