+++ lib/mix/lib/mix/tasks/profile.fprof.ex
@@ -102,14 +102,14 @@ defmodule Mix.Tasks.Profile.Fprof do
   this should give you a more correct insight into your real bottlenecks.
   Profiling with other environments might produce some false bottlenecks, such as
   protocol dispatches, which perform much faster with the `prod` environment when
-  `build_embedded` is `true` (which is the default for production).
+  `:build_embedded` is `true` (which is the default for production).
   """
 
   @spec run(OptionParser.argv) :: :ok
   def run(args) do
     unless Mix.Project.config[:build_embedded] do
-      Mix.shell.error "Warning: It's advised to run this task when build_embedded is set " <>
-                      "to true (usually the prod environment). Otherwise the results may " <>
+      Mix.shell.error "Warning: It's advised to run this task when :build_embedded is set " <>
+                      "to \"true\" (usually the prod environment). Otherwise the results may " <>
                       "contain false bottlenecks which will not appear in production."
     end
 
@@ -259,7 +259,7 @@ defmodule Mix.Tasks.Profile.Fprof do
     print_row(["s", "B", ".3f", ".3f", "s"], ["Total", count, acc, own, ""])
   end
 
-  # Represents the `pid` entry
+  # Represents the "pid" entry
   defp print_analysis_result([{pid_atom, count, :undefined, own} | info]) do
     print_process(pid_atom, count, own)